operation/pyproject.toml
[project]
name = "github-actions-self-hosted-runner-container"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Atsushi Morimoto (74th)", email = "74th.tech@gmail.com" }
]
dependencies = [
    "pygithub>=2.3.0",
]
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "httpx>=0.27.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/github_actions_self_hosted_runner_container"]
